pyproject.toml
[project]
name = "fiftwall-test-case"
version = "0.1.0"
description = ""
authors = [
    {name = "Manuel Diaz",email = "alejandromadr@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "graphviz (>=0.20.3,<0.21.0)",
    "langchain (>=0.3.22,<0.4.0)",
    "langchain-community (>=0.3.20,<0.4.0)",
    "langchain-core (>=0.3.49,<0.4.0)",
    "langgraph (>=0.3.21,<0.4.0)",
    "langchain-openai (>=0.3.11,<0.4.0)",
    "pydantic (>=2.11.1,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "supabase (>=2.15.0,<3.0.0)",
    "datasets (>=3.5.0,<4.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "openai (>=1.70.0,<2.0.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "tenacity (>=9.0.0,<10.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "chromadb (>=0.6.3,<0.7.0)",
    "langchain-chroma (>=0.2.2,<0.3.0)",
    "requests (>=2.32.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
